optimize zoom camera

+++ ui/src/main/java/com/arjinmc/smartcam/ui/activity/SmartCamComplexActivity.java
@@ -202,7 +202,7 @@ public class SmartCamComplexActivity extends AppCompatActivity implements View.O
 
         if (mSmartCam.isZoomAvailable()) {
             mSbZoom.setVisibility(View.VISIBLE);
-            mSbZoom.setMax((int) mSmartCam.getMaxZoom());
+            mSbZoom.setMax(mSmartCam.getMaxZoom());
         }
     }
